Update rancher release for new dir structure
Rancher charts dir structure changed to have separate directory for each
releases. This change modifies the release script to create the
appropriate dir structure and also sign-off the commits.

The requirements include, creating a base for the new release using the
existing latest release and committing it separately, and then adding
the changes for new release on top of the copied base as another commit.